+++ bot_imports.py
@@ -1,0 +1,2 @@
+print("Merlin bot written in python by windowsboy111 :)")
+print('==> Importing libraries...')